power-policy: Remove automatic provider upgrade logic (#293)
Remove logic to automatically upgrade a provider contract if power is
available. This can result in a device receiving more power than needed
while another device receives less. This is particularly a concern for
TBT compliance testing. With this change devices will have to explicitly
request a higher contract. This also removes the recovery logic which
was only used in cases where we attempted to downgrade a contract. The
new behavior does not attempt to downgrade contracts, so it's no longer
needed.