Update CODEOWNERS to use EC team (#330)
With growing number of repositories, having consistent and manageable
list of codeowners becomes important. This PR updates the CODEOWNERS
file to use the EC code owners team as the default codeowners for the
repository.
This hardens the repository against stale permission issues down the
line by tying the access to being part of the team rather than being
explicitly granted permission to a repository.

This PR also adds `--locked` to a workflow check that was missed in the
previous PRs.