Add SyncCell type to support Sync fix in embassy upstream (#365)
Embassy upstream recently fixed accidentally adding Sync to non-Sync
types: https://github.com/embassy-rs/embassy/pull/4320

In preparation to fix #361, added a drop-in replacement for Cell where
moving to a proper primitive usage such as embassy_sync::mutex::Mutex is
untenable. This also refactors the intrusive_list to utilize this new
SyncCell type, allowing it to be used directly without any containers.

Note that in most cases SyncCell is not the best solution. If it is
possible, proper synchronization primitives should be leveraged.

A critical section was used as opposed to a blocking_mutex type, as
blocking_mutex is non-re-entrant. This allows the most flexibility in
supporting scenarios with SyncCell, and does not require permutated get
and set implementations based on mutex backing type.

In short:
- New SyncCell type introduced to unblock ingesting embassy upstream
Sync fix
- We should use it sparingly
- IntrusiveList was updated to use this new type, allowing it to be
allocated statically without any errors

---------

Co-authored-by: Melvin Wang <melvin.mc.wang@gmail.com>