Only send `SRDY` after Sink Ready interrupt (#375)
Previously, we sent `SRDY` when we received a new consumer contract, but
this may fail if PD controller isn't ready to sink power from the
provider. The controller indicates this with the Sink Ready interrupt,
which is fired when an explicit or implicit contract has been
established. As such, Sink Ready always implies a contract is available,
and `type_c_service::driver::tps6699x::Tps6699x::update_port_status()`
ensures that `PortStatus::available_sink_contract` is valid by the time
that `type_c_service::wrapper::ControllerWrapper::process_event()` is
called.

Now, we only notify power policy of the new sink contract after the Sink
Ready interrupt. In turn, that notification becomes a request to send
`SRDY` to the PD controller.

Fixes #338 

Also fixes a bug where we were incorrectly recovering a
`power::policy::device::Device` found in an incorrect state.

---------

Co-authored-by: Jerry Xie <139205137+jerrysxie@users.noreply.github.com>