Add ability to send PollStaticData after initializaiton, and add unique return type for when Timeout recovery fails (#381)
In #367, the ability for the battery state machine to autonomously move states was removed. Currently, there is no way to get out of the `State::Present(PresentSubstate::Operational(OperationalSubstate::Init))` state, which is hit after initialization. Sending a `PollStaticData` cmd should get you into the polling state, and this change makes it so you can send a `PollStaticData` cmd to get to the polling state.