Implement ThreadModeCell (#380)
This PR introduces `ThreadModeCell`, a lighter-weight alternative to
`SyncCell` which implements `Sync` for `Cell` but is only safe on
single-core systems. Instead of holding a critical section, this simply
asserts that we are in thread mode before accessing the inner `Cell`,
which prevents the possibility of data races if there are attempts to
access it within an interrupt (instead panicking).

This essentially follows the same pattern as a `ThreadModeRawMutex`, but
gives us the `Clone` behavior of a `Cell` and can be used more easily in
a sync context.

`SyncCell` is now aliased to either `CriticalSectionCell` or
`ThreadModeCell` depending on environment.