Remove update from ThreadModeCell (#404)
The `update()` method on `Cell` wasn't stabilized until 1.88, but our
current MSRV is 1.85. This was not caught by our workflow as
`ThreadModeCell` is only built when target is thumbv8m, but our MSRV
workflow doesn't specifically check that target. A new PR will also be
made to update our workflow to catch this next time.

This is technically a breaking change but AFAIK no one is using the
update method currently.