Type-c: Event loop refactoring (#390)
This PR is a refactoring of the type-C event loop to achieve the
following:

* Open up the wrapper event loop to allow for users of the code to
control event processing. This moves the wrapper code over to use a
stream of port events. This better matches how the type-C service
implementation events are.
* Add support for alert/VDM. The existing event bitflags are now split
up. The existing port status related flags are in their own portion and
the alert and VDM ones are in their own. This is meant to reflect how
they're used in code. The status flags are related to port state and are
generally handled by the same code paths. While the alert and VDM events
are notifications to us that are don't really impact the port state and
will be handled by completely different code.

Also increase the delays in one of the std examples, 250 ms is very
short to simulate a plug and unplug.