Implement initial thermal service (#383)
This overhauls the first attempt of a thermal service to be more
flexible. The thermal service provides:

- Async friendly approach to interfacing with sensors and fans. Each
sensor and fan runs in its own task, waiting to receive and process
commands and performing additional duties described below.
- Each sensor is responsible for periodically sampling and caching
temperature samples to be quickly retrieved when asked.
- Each sensor is responsible for checking if the most recent temperature
sample exceeds several thresholds (Warn, Prochot, Critical) and
generating an event if so.
- These events can be listened for by OEM to handle appropriately (such
as throttling AP in case of Prochot)
- Sensors support various requests such as changing sampling period (to
sample more frequently as temperature rises) or to be disabled when
needed to save power
- Sensors support changing thermal profile at runtime to more easily
change behavior at runtime
- Each fan supports independent auto control by querying an associated
sensor when it wants to update its speed
- Fans can be set to manual control (as an example, to enable a passive
cooling policy where AP is throttled instead of turning fans on
- Fans also support multiple profiles such as sensors
- A default linear response is provided when the fan is in the ramping
state, however OEM can easily plug in their own custom control algorithm
instead
- An MPTF handler which can translate MPTF thermal commands according to
spec to thermal service functions

This design should strike a fair balance between flexibility and
providing usefulness. It is continually being tweaked (especially
implementation details) so this PR is currently a draft.

Resolves #132 
Resolves #133 
Resolves #134 
Resolves #135 
Resolves #136 
Resolves #137 
Resolves #222

---------

Co-authored-by: Jerry Xie <jerryxie@microsoft.com>