type-c-service/wrapper: Remove port count generics (#497)
This PR refactors the backing for the controller wrapper with the goal
of removing port count generic arguments. The existing backing types
that were used to supply certain data structures have been expanded to
also include all needed mutable state and service registration objects.
This also helps to simplify constructing wrappers a bit. This PR
introduces the types `Storage`, `ReferencedStorage`, and `Backing` to
accomplish this task. This PR also introduces a `DynPortState` trait
that is used to erase the port count generic from the wrapper.

Since the wrapper no longer has to be monomorphized per port count this
results in a size reduction of at least ~14K per monomorphized copy
removed.

This PR also includes some general clean-up, most notable is that the
wrapper now accepts a `M: RawMutex` argument to allow users of this type
to the supply the mutex type.