Migrate ODP MCTP message types from mctp-rs (#513)
Currently, the definitions for all the messages we want to send over
MCTP are in the mctp-rs crate protected by a feature flag.
These messages are tightly coupled with embedded-services, so it doesn't
really make sense to have them over there. This change migrates those
message types over here so we can iterate on them and clients of mctp-rs
don't need to deal with implementation details of embedded-services.

In the course of porting, discovered and fixed a bug that would cause deserialization of the BatteryGetBstResponse message to yield an incorrect result.