Remove dependency on `embassy-executor` (#594)
All of our services should be able to run agnostic of an async executor.
While we use embassy, the ability to use tokio for testing in std
environments should be supported.

All tasks that should be spawned in a thread have been moved to each
subsystem's respective `task` namespace, so that users know to spawn all
async functions within that namespace.