Remove unnecessary parameters unconstrained_power from Func cmp_consumer_capability (#636)
In current ODP code
[cmp_consumer_capability()](https://github.com/OpenDevicePartnership/embedded-services/blob/2eaac0fdddf76e52fab5f24cffc635c69037de94/power-policy-service/src/consumer.rs#L25),
when two power sources have the same power capability, it will compare
their `unconstrained_power` next.
Currently, the unconstrained power of USBC power is always false, so if
the unconstrained power of another lower-power source is always true, it
will always select this lower-power source regardless of whether it's
inserted first or last. It is enough to use power capability and current
power flag to choose best consumer.