keyboard-service: fix syntax error in task macro (#640)
#594 moved the responsibility for declaring async tasks from
embedded-services to the application layer. To support this, some tasks
that needed to be generic over a type (which is not directly supported
by embassy) made into macros that emit a function that uses a concrete
type. The application layer is required to call that macro and then
implement a task that invokes the generated function.

However, in the case of the keyboard service, there appears to be a typo
causing the name of the type identifier being passed into the macro to
not match the name of the type identifier that is actually used in the
macro, resulting in a compilation error whenever a user tries to
actually invoke the macro. This change fixes the typo.